@@ -12,37 +12,37 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/platform-server": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@ngx-translate/core": "^8.0.0",
-    "@ngx-translate/http-loader": "^2.0.0",
-    "ace-builds": "^1.2.6",
-    "angular-sortablejs": "^2.5.1",
+    "@angular/animations": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/platform-server": "6.0.0",
+    "@angular/router": "6.0.0",
+    "@ngx-translate/core": "^10.0.1",
+    "@ngx-translate/http-loader": "^3.0.1",
+    "ace-builds": "^1.3.3",
+    "angular-sortablejs": "^2.5.2",
     "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.6",
     "dropzone": "^5.3.0",
-    "hashedit": "^0.6.30",
-    "moment": "^2.18.0",
-    "rxjs": "^5.1.0",
+    "moment": "^2.22.1",
+    "rxjs": "^6.1.0",
     "sortablejs": "^1.7.0",
-    "typescript": "^2.2.1",
+    "typescript": "2.7.2",
     "web-animations-js": "^2.2.2",
-    "zone.js": "^0.7.6"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.0",
-    "@angular/compiler-cli": "^2.4.0",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
     "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
+    "@types/node": "^6.0.108",
     "codelyzer": "~2.0.0",
     "jasmine-core": "~2.5.2",
     "jasmine-spec-reporter": "~3.2.0",
@@ -50,12 +50,12 @@
     "karma-chrome-launcher": "~2.0.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "~1.1.0",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.0.0"
+    "typescript": "2.7.2"
   },
   "description": "Respond 6 is an open source, responsive content management system built on the Lumen PHP and Angular (currently 4) frameworks. Respond 6 is built so that you can quickly build beautiful, responsive websites. The latest version features flat-file storage, a beautiful visual editor, and fast static sites. Learn more about Respond at https://respondcms.com.",
   "main": "karma.conf.js",
