@@ -62,7 +62,7 @@ export class RemoveProductComponent {
 
     this._productService.remove(this.model.id)
                      .subscribe(
-                       data => { this.success(); },
+                      (data: any) => { this.success(); },
                        error =>  { this.onError.emit(<any>error); }
                       );
 
