@@ -90,7 +90,7 @@ export class DrawerComponent {
     // retrieve settings
     this._appService.retrieveSettings()
                      .subscribe(
-                       data => {
+                      (data: any) => {
                          this.siteUrl = data.siteUrl;
                          this.siteUrl = this.siteUrl.replace('{{siteId}}', this.id);
                        },
