@@ -67,7 +67,7 @@ export class RemoveFileComponent {
 
     this._fileService.remove(this.model.name)
                      .subscribe(
-                       data => { this.success(); },
+                       (data: any) => { this.success(); },
                        error => { this.onError.emit(<any>error) }
                       );
 
