@@ -63,7 +63,7 @@ export class AddMenuItemComponent {
     // list pages
     this._pageService.list()
                      .subscribe(
-                       data => { this.pages = data; },
+                      (data: any) => { this.pages = data; },
                        error =>  this.errorMessage = <any>error
                       );
 
@@ -84,7 +84,7 @@ export class AddMenuItemComponent {
 
     this._menuItemService.add(this.menu.id, this.model.html, this.model.cssClass, this.model.isNested, this.model.url, this.model.target)
                      .subscribe(
-                       data => { this.success(); },
+                      (data: any) => { this.success(); },
                        error =>  { this.onError.emit(<any>error); }
                       );
 
