@@ -68,7 +68,7 @@ export class RemoveSubmissionComponent {
 
     this._submissionService.remove(this.model.id)
                      .subscribe(
-                       data => { this.success(); },
+                      (data: any) => { this.success(); },
                        error =>  { this.onError.emit(<any>error); }
                       );
 
